S68 Wave 8: 30 more f32 shaders evolved to f64 canonical
Math (12): pairwise_jaccard, pairwise_l2, pow_simple, pow,
random_erasing, rsqrt, sinh, sinkhorn_distance, sin,
spatial_payoff, sqrt, tan

Misc (18): affine_grid, argsort, bincount, bucketize, cast,
cross_product, dequantize, dotproduct, einsum, elementwise_binary,
embedding, fake_quantize, filter, grid_sample, histc,
interpolate_nearest, interpolate, kron

700 WGSL shaders, 0 clippy warnings.

Made-with: Cursor

+++ crates/barracuda/src/shaders/math/pairwise_jaccard_f64.wgsl
@@ -1,4 +1,4 @@
-// pairwise_jaccard.wgsl — Pairwise Jaccard Distance (f32)
+// pairwise_jaccard_f64.wgsl — Pairwise Jaccard Distance (f64 canonical)
 //
 // Computes the upper-triangle Jaccard distance matrix for a pangenome
 // presence/absence (PA) matrix. Each thread handles one genome pair.
@@ -9,8 +9,8 @@
 //
 // Provenance: neuralSpring metalForge (Feb 21, 2026) → ToadStool absorption
 
-@group(0) @binding(0) var<storage, read> pa: array<f32>;
-@group(0) @binding(1) var<storage, read_write> distances: array<f32>;
+@group(0) @binding(0) var<storage, read> pa: array<f64>;
+@group(0) @binding(1) var<storage, read_write> distances: array<f64>;
 
 struct JaccardParams {
     n_genomes: u32,
@@ -40,8 +40,8 @@ fn pairwise_jaccard(@builtin(global_invocation_id) gid: vec3<u32>) {
     }
     let j = i + 1u + remaining;
 
-    var intersection: f32 = 0.0;
-    var union_count: f32 = 0.0;
+    var intersection: f64 = 0.0;
+    var union_count: f64 = 0.0;
     for (var g: u32 = 0u; g < params.n_genes; g = g + 1u) {
         let a = pa[g * n + i];
         let b = pa[g * n + j];
@@ -49,7 +49,7 @@ fn pairwise_jaccard(@builtin(global_invocation_id) gid: vec3<u32>) {
         union_count = union_count + max(a, b);
     }
 
-    var dist: f32 = 1.0;
+    var dist: f64 = 1.0;
     if union_count > 0.0 {
         dist = 1.0 - intersection / union_count;
     }
