+++ crates/barracuda/src/shaders/misc/filter_f64.wgsl
@@ -1,42 +1,33 @@
-// filter.wgsl — Stream compaction (predicate evaluation + scatter)
+// filter_f64.wgsl — Stream compaction (predicate evaluation + scatter) (f64 canonical)
 //
 // Pass 1 — evaluate_predicate:
 //   flags[i] = 1 if input[i] satisfies the predicate, else 0.
 //
-// (Passes 2a/2b run the prefix_sum.wgsl shader — see prefix_sum.wgsl)
-//   After them, scan[i] = exclusive prefix sum of flags.
-//
 // Pass 3 — scatter:
 //   output[scan[i]] = input[i] if flags[i] == 1
-//   total[0]        = number of selected elements (scan[N-1] + flags[N-1])
 //
 // Operations:
-//   0 = GreaterThan    (value > threshold)
-//   1 = LessThan       (value < threshold)
-//   2 = Equal          (|value - threshold| < ε)
-//   3 = NotEqual       (|value - threshold| ≥ ε)
-//   4 = GreaterOrEqual (value >= threshold)
-//   5 = LessOrEqual    (value <= threshold)
+//   0 = GreaterThan, 1 = LessThan, 2 = Equal, 3 = NotEqual,
+//   4 = GreaterOrEqual, 5 = LessOrEqual
 
 struct FilterParams {
     size:      u32,
     operation: u32,
-    n_groups:  u32,   // ceil(size / 256) — for pass 2 dispatch
+    n_groups:  u32,
     _pad:      u32,
-    threshold: f32,
-    epsilon:   f32,   // equality tolerance (default 1e-5)
-    _pad2:     f32,
-    _pad3:     f32,
+    threshold: f64,
+    epsilon:   f64,
+    _pad2:     f64,
+    _pad3:     f64,
 }
 
-@group(0) @binding(0) var<storage, read>       input:   array<f32>;
+@group(0) @binding(0) var<storage, read>       input:   array<f64>;
 @group(0) @binding(1) var<storage, read_write> flags:   array<u32>;
 @group(0) @binding(2) var<storage, read_write> scan:    array<u32>;
-@group(0) @binding(3) var<storage, read_write> output:  array<f32>;
-@group(0) @binding(4) var<storage, read_write> total:   array<u32>; // [0] = count
+@group(0) @binding(3) var<storage, read_write> output:  array<f64>;
+@group(0) @binding(4) var<storage, read_write> total:   array<u32>;
 @group(0) @binding(5) var<uniform>             params:  FilterParams;
 
-// ── Pass 1: evaluate predicate ───────────────────────────────────────────────
 @compute @workgroup_size(256)
 fn evaluate_predicate(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let gid = global_id.x;
@@ -60,7 +51,6 @@ fn evaluate_predicate(@builtin(global_invocation_id) global_id: vec3<u32>) {
     flags[gid] = select(0u, 1u, keep);
 }
 
-// ── Pass 3: scatter selected elements using prefix-sum indices ────────────────
 @compute @workgroup_size(256)
 fn scatter(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let gid = global_id.x;
@@ -70,7 +60,6 @@ fn scatter(@builtin(global_invocation_id) global_id: vec3<u32>) {
         output[scan[gid]] = input[gid];
     }
 
-    // Thread 0 of the last workgroup writes the total count.
     let last = params.size - 1u;
     if (gid == last) {
         total[0] = scan[last] + flags[last];
